kvm.sh symlink support
if mono exists at the kvm.sh folder, it will be symlinked in the unpacked kre folder. this enables kvm.sh to be tightly associated with a specific build on mono on the machine.

kvm.sh will symlink a folder named current to the most recently 'use'd kre/bin version. this enables a symlink link to current/k to pass through on the machine