Fixing kvm-Global-upgrade
When kvm upgrade is executed with a -g flag, today the code tries to fetch the latest version details from the REST url passing in the selectedRuntime and selectedArch explicitly.

These parameters have null values when this command is executed. Right thing is to have code exactly similar to Kvm-Upgrade but pass on global flag to Kvm-Install. This also avoid duplicate code on (Needs-Elevation) case. I veriried this fixes the issue.