Prepare dnvm.sh for CoreCLR on Linux/OSX
dnvm.sh was unable to handle cases where a runtime package had an os and
architecture included in the name.  For example, using the package name:

  dnx-coreclr-linux-x64.1.0.0-dev

Exposed some bugs. For example, setting an alias for a runtime and then
trying to set that runtime using the alias didn't work.

Fix these issues as well as include the architecture in the output in
dnvm list.  While the CoreCLR team will only be producing x64 of the
runtime in the short term, it will be good to set up the infrastructure
for the day new architectures come online.