Add runtime and arch flags to use command
Previously, you had to use a full package name to select a coreclr based
runtime with dnvm use.  With this change we now support the -r and -a
flags line in powershell.  For example, the following all work:

 $ dnvm use 1.0.0-dev -r coreclr
 $ dnvm use 1.0.0-dev -r mono
 $ dnvm use 1.0.0-dev -r coreclr -a x64
 $ dnvm use 1.0.0-dev -r coreclr -a x86

For mono based runtimes we ignore the architecture if one is passed in
on the command line.  We defaut to x64 as an architecture for coreclr if
one is not set explicitly.